feat: Implement ViewModel and ViewModelFactory on Activity
