+++ app/src/main/java/com/android/example/music/MainActivity.kt
@@ -2,30 +2,25 @@ package com.android.example.music
 
 import android.Manifest
 import android.content.pm.PackageManager
-import android.media.MediaMetadataRetriever
-import android.net.Uri
 import android.os.Build
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.Environment
-import android.util.Log
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-import com.android.example.music.models.Song
-import com.android.example.music.player.MusicPlayer
-import com.android.example.music.player.MusicPlayerImplementation
-import java.io.File
+import androidx.lifecycle.ViewModelProvider
 
 class MainActivity : AppCompatActivity() {
-    lateinit var musicPlayer: MusicPlayer
-
+    lateinit var viewModelFactory: MainActivityViewModelFactory
+    lateinit var viewModel: MainActivityViewModel
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
+        viewModelFactory = MainActivityViewModelFactory(application)
+        viewModel =
+            ViewModelProvider(this, viewModelFactory)[MainActivityViewModel::class.java]
         if (allPermissionsGranted()) {
-            initializePlayer()
+            viewModel.initializePlayer()
         } else {
             ActivityCompat.requestPermissions(
                 this@MainActivity, REQUIRED_PERMISSIONS, REQUEST_CODE_PERMISSIONS
@@ -46,7 +41,7 @@ class MainActivity : AppCompatActivity() {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         if (requestCode == REQUEST_CODE_PERMISSIONS) {
             if (allPermissionsGranted()) {
-                initializePlayer()
+                viewModel.initializePlayer()
             } else {
                 Toast.makeText(
                     applicationContext,
@@ -58,25 +53,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun initializePlayer() {
-        val folder = Environment.getExternalStorageDirectory()
-        val songsList = File(folder, CHILD_ROUTE)
-            .listFiles()?.mapIndexed { index, item ->
-                val metadataRetriever = MediaMetadataRetriever()
-                val uri = Uri.parse(item.path)
-                metadataRetriever.setDataSource(this, uri)
-                val trackName =
-                    metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_TITLE)
-                        .toString()
-                Song(path = item.path, index = index, name = trackName)
-            }
-        songsList?.let { list ->
-            musicPlayer = MusicPlayerImplementation(list)
-        }
-    }
-
     companion object {
-        private const val CHILD_ROUTE = "Music/Music-App"
         private const val REQUEST_CODE_PERMISSIONS = 100
         private val REQUIRED_PERMISSIONS =
             mutableListOf<String>().apply {
