+++ app/src/main/java/com/android/example/music/MainActivityViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.android.example.music
+
+import android.app.Application
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class MainActivityViewModelFactory(private val application: Application) :
+    ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainActivityViewModel::class.java)) {
+            return MainActivityViewModel(application) as T
+        }
+        return super.create(modelClass)
+    }
+}
